+++ segtypes/address_range.py
@@ -1,0 +1,9 @@
+from dataclasses import dataclass
+
+@dataclass
+class AddressRange:
+    start: int
+    end: int
+    
+    def contains(self, addr: int) -> bool:
+        return self.start <= addr < self.end
